Add files via upload
Um, I couldn't get the Random# game. I DID have a patch-y one that kinda-sorta worked (it could be played!), but of course I left it on my lab computer, so I can't upload it here (@home). So, I appropriated the code from my teammate. I am honest: this code isn't mine, it is Wai's. He is a talented coder. I am trying to learn/get it, so don't count me out. 

But that random number thing is difficult. AND I can't install Python here. And if I accidentally commence an endless loop in repl.it my computer goes berserk...